+++ src/Services/IM/IMServiceBuilder.php
@@ -24,6 +24,7 @@ use Bitrix24\SDK\Services\IM\Notify\Service\Notify;
 use Bitrix24\SDK\Services\IM\Placements\PlacementLocationCodes;
 use Bitrix24\SDK\Services\IM\Placements\Placements;
 use Bitrix24\SDK\Services\IM\Revision\Service\Revision;
+use Bitrix24\SDK\Services\IM\User\Service\UserStatus;
 use Bitrix24\SDK\Services\Placement\Service\Placement;
 
 #[ApiServiceBuilderMetadata(new Scope(['im']))]
@@ -78,6 +79,15 @@ class IMServiceBuilder extends AbstractServiceBuilder
     {
         if (!isset($this->serviceCache[__METHOD__])) {
             $this->serviceCache[__METHOD__] = new Counters($this->core, $this->log);
+        }
+
+        return $this->serviceCache[__METHOD__];
+    }
+
+    public function userStatus(): UserStatus
+    {
+        if (!isset($this->serviceCache[__METHOD__])) {
+            $this->serviceCache[__METHOD__] = new UserStatus($this->core, $this->log);
         }
 
         return $this->serviceCache[__METHOD__];