+++ tests/Unit/Services/IM/IMServiceBuilderTest.php
@@ -17,6 +17,7 @@ use Bitrix24\SDK\Services\IM\Counters\Service\Counters;
 use Bitrix24\SDK\Services\IM\Dialog\Service\Dialog;
 use Bitrix24\SDK\Services\IM\IMServiceBuilder;
 use Bitrix24\SDK\Services\IM\Placements\Placements;
+use Bitrix24\SDK\Services\IM\User\Service\UserStatus;
 use Bitrix24\SDK\Services\ServiceBuilder;
 use Bitrix24\SDK\Tests\Unit\Stubs\NullBatch;
 use Bitrix24\SDK\Tests\Unit\Stubs\NullBulkItemsReader;
@@ -60,6 +61,12 @@ class IMServiceBuilderTest extends TestCase
     {
         $this->assertInstanceOf(Counters::class, $this->serviceBuilder->counters());
         $this->assertSame($this->serviceBuilder->counters(), $this->serviceBuilder->counters());
+    }
+
+    public function testGetUserStatusService(): void
+    {
+        $this->assertInstanceOf(UserStatus::class, $this->serviceBuilder->userStatus());
+        $this->assertSame($this->serviceBuilder->userStatus(), $this->serviceBuilder->userStatus());
     }
 
     public function testGetPlacementsService(): void